change cost matrix to floats and fix tests

+++ local_planner/src/nodes/candidate_direction.h
@@ -3,11 +3,11 @@
 namespace avoidance {
 
 struct candidateDirection {
-  double cost;
-  double elevation_angle;
-  double azimuth_angle;
+  float cost;
+  float elevation_angle;
+  float azimuth_angle;
 
-  candidateDirection(double c, double e, double z)
+  candidateDirection(float c, float e, float z)
       : cost(c), elevation_angle(e), azimuth_angle(z){};
 
   bool operator<(const candidateDirection& y) const { return cost < y.cost; }
